requirements.txt
torch
spacy
gradios